# tripmatch_backend/docs/api-contract-iteratia-1-2.openapi.yaml
openapi: 3.0.3
info:
  title: TripMatch Backend API Contract (Iteratia 1 + 2 + 3)
  version: "iter1-iter2-iter3"
  description: >
    Contract extras direct din codebase (controller + DTO + validari service).
    Toate endpointurile protejate folosesc Basic Auth, exceptand /api/auth/**.
servers:
  - url: /
security:
  - basicAuth: []
...
paths:
...
  /api/admin/accounts:
    get:
      tags: [Admin]
      summary: List accounts (admin only)
      parameters:
        - name: isActive
          in: query
          required: false
          schema:
            type: boolean
        - name: q
          in: query
          required: false
          schema:
            type: string
          description: Search text; blank string is treated as null.
        - name: role
          in: query
          required: false
          schema:
            type: string
            enum: [USER, AGENCY, ADMIN]
          description: Case-insensitive. Any other value -> 400 "Invalid role filter".
        - $ref: "#/components/parameters/pageParamDefault0"
        - $ref: "#/components/parameters/sizeParamDefault20"
        - $ref: "#/components/parameters/sortParamAdminAccounts"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageAdminAccountListItemResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/admin/accounts/{accountId}/block:
    parameters:
      - $ref: "#/components/parameters/positiveAccountIdPath"
    patch:
      tags: [Admin]
      summary: Block account (admin only)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlockAccountResponse"
              example:
                message: "Account blocked successfully!"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/admin/accounts/{accountId}/unblock:
    parameters:
      - $ref: "#/components/parameters/positiveAccountIdPath"
    patch:
      tags: [Admin]
      summary: Unblock account (admin only)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnblockAccountResponse"
              example:
                message: "Account unblocked successfully!"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/requests/{requestId}/offers:
    parameters:
      - $ref: "#/components/parameters/positiveRequestIdPath"
    post:
      tags: [Offer]
      summary: Create offer on a published request (agency only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOfferRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OfferResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/offers/{offerId}:
    parameters:
      - $ref: "#/components/parameters/positiveOfferIdPath"
    get:
      tags: [Offer]
      summary: Get offer details (agency owner or user owner via request)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OfferDetailResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Offer]
      summary: Update offer (agency owner)
      description: >
        Permis doar daca oferta curenta este DRAFT sau READY.
        In request, status poate fi doar DRAFT sau READY.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOfferRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OfferResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Offer]
      summary: Delete offer (agency owner)
      description: Permis doar pentru status DRAFT sau READY.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OfferResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/offers/{offerId}/send:
    parameters:
      - $ref: "#/components/parameters/positiveOfferIdPath"
    post:
      tags: [Offer]
      summary: Send offer (agency owner)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OfferResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/offers/{offerId}/accept:
    parameters:
      - $ref: "#/components/parameters/positiveOfferIdPath"
    post:
      tags: [Offer]
      summary: Accept offer (user owner)
      description: >
        Endpoint protejat contra race condition prin locking pesimist pe ofertele aceluiasi request.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OfferResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/offers/{offerId}/reject:
    parameters:
      - $ref: "#/components/parameters/positiveOfferIdPath"
    post:
      tags: [Offer]
      summary: Reject offer (user owner)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectOfferRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OfferResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/offers/agency:
    get:
      tags: [Offer]
      summary: List offers created by current agency
      parameters:
        - $ref: "#/components/parameters/pageParamDefault0"
        - $ref: "#/components/parameters/sizeParamDefault20"
        - $ref: "#/components/parameters/sortParamOffers"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageOfferListItemResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/offers/user:
    get:
      tags: [Offer]
      summary: List offers received by current user (through own requests)
      parameters:
        - $ref: "#/components/parameters/pageParamDefault0"
        - $ref: "#/components/parameters/sizeParamDefault20"
        - $ref: "#/components/parameters/sortParamOffers"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageOfferListItemResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/notifications:
    get:
      tags: [Notification]
      summary: List my notifications
      parameters:
        - $ref: "#/components/parameters/pageParamDefault0"
        - $ref: "#/components/parameters/sizeParamDefault20"
        - $ref: "#/components/parameters/sortParamNotifications"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageNotificationListItemResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/notifications/unread-count:
    get:
      tags: [Notification]
      summary: Get unread notifications count
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnreadNotificationsCountResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/notifications/{notificationId}/read:
    parameters:
      - $ref: "#/components/parameters/positiveNotificationIdPath"
    patch:
      tags: [Notification]
      summary: Mark single notification as read
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/notifications/read:
    patch:
      tags: [Notification]
      summary: Mark selected notifications as read
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarkNotificationsReadRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/notifications/read-all:
    patch:
      tags: [Notification]
      summary: Mark all notifications as read
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  parameters:
    pageParamDefault0:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
...
    sizeParamDefault20:
      name: size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 20
...
    sortParamOffers:
      name: sort
      in: query
      required: false
      schema:
        type: string
        default: "createdAt,desc"
      description: Spring pageable sort, ex `createdAt,desc`.
    sortParamNotifications:
      name: sort
      in: query
      required: false
      schema:
        type: string
        default: "createdAt,desc"
      description: Spring pageable sort, ex `createdAt,desc`.
    sortParamAdminAccounts:
      name: sort
      in: query
      required: false
      schema:
        type: string
        default: "createdAt,desc"
      description: Spring pageable sort, ex `createdAt,desc`.
...
    positiveAccountIdPath:
      name: accountId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
...
    positiveRequestIdPath:
      name: requestId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
    positiveOfferIdPath:
      name: offerId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
    positiveNotificationIdPath:
      name: notificationId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
...
  responses:
    BadRequest:
      description: >
        Bad Request. In acest backend, body-ul poate fi:
        1) string simplu (Business error) sau
        2) array de stringuri (Bean validation errors).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationOrMessageError"
    Unauthorized:
      description: Missing/invalid Basic Auth credentials.
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            type: string
          example: "You cannot edit another user's profile"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            type: string
          example: "Trip not found!"
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            type: string
          example: "Activity overlaps with existing activities: [11, 15]. Retry with allowOverlap=true if intentional."
...
  schemas:
    ValidationOrMessageError:
      oneOf:
        - type: string
        - type: array
          items:
            type: string
...
    AdminAccountListItemResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        email: { type: string }
        role: { type: string }
        isActive: { type: boolean }
        isApproved: { type: boolean }
        firstName: { type: string, nullable: true }
        lastName: { type: string, nullable: true }
        agencyName: { type: string, nullable: true }
        createdAt: { type: string, format: date-time, nullable: true }
        lastLoginAt: { type: string, format: date-time, nullable: true }
...
    BlockAccountResponse:
      type: object
      properties:
        message: { type: string }
      required: [message]
    UnblockAccountResponse:
      type: object
      properties:
        message: { type: string }
      required: [message]
...
    PageAdminAccountListItemResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/AdminAccountListItemResponse"
        page: { type: integer }
        size: { type: integer }
        totalElements: { type: integer, format: int64 }
        totalPages: { type: integer }
        last: { type: boolean }
      required: [content, page, size, totalElements, totalPages, last]
...
    OfferStatus:
      type: string
      enum: [DRAFT, READY, SENT, ACCEPTED, REJECTED, EXPIRED]
    NotificationType:
      type: string
      enum: [OFFER_SENT, OFFER_ACCEPTED, OFFER_REJECTED]
...
    CreateOfferRequest:
      type: object
      properties:
        price:
          type: number
          minimum: 0.01
        currency:
          type: string
          maxLength: 64
          pattern: "^.*\\S.*$"
        accommodationDetails:
          type: string
          maxLength: 512
          pattern: "^.*\\S.*$"
        transportDetails:
          type: string
          maxLength: 512
          pattern: "^.*\\S.*$"
        itinerarySummary:
          type: string
          nullable: true
        conditions:
          type: string
          nullable: true
        attachmentUrl:
          type: string
          maxLength: 1024
          nullable: true
      required: [price, currency, accommodationDetails, transportDetails]
    UpdateOfferRequest:
      type: object
      description: At least one field must be provided.
      properties:
        price:
          type: number
          minimum: 0.01
        currency:
          type: string
          maxLength: 64
        accommodationDetails:
          type: string
          maxLength: 512
        transportDetails:
          type: string
          maxLength: 512
        itinerarySummary:
          type: string
        conditions:
          type: string
        attachmentUrl:
          type: string
          maxLength: 1024
        status:
          $ref: "#/components/schemas/OfferStatus"
    RejectOfferRequest:
      type: object
      properties:
        feedbackMessage:
          type: string
          maxLength: 4000
          pattern: "^.*\\S.*$"
      required: [feedbackMessage]
    OfferResponse:
      type: object
      properties:
        offerId: { type: integer, format: int64 }
        status:
          $ref: "#/components/schemas/OfferStatus"
        message: { type: string }
      required: [offerId, status, message]
    OfferListItemResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        requestId: { type: integer, format: int64 }
        price: { type: number }
        currency: { type: string }
        status:
          $ref: "#/components/schemas/OfferStatus"
        sentAt: { type: string, format: date-time, nullable: true }
        respondedAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    OfferDetailResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        requestId: { type: integer, format: int64 }
        agencyAccountId: { type: integer, format: int64 }
        price: { type: number }
        currency: { type: string }
        accommodationDetails: { type: string }
        transportDetails: { type: string }
        itinerarySummary: { type: string, nullable: true }
        conditions: { type: string, nullable: true }
        attachmentUrl: { type: string, nullable: true }
        feedbackMessage: { type: string, nullable: true }
        status:
          $ref: "#/components/schemas/OfferStatus"
        sentAt: { type: string, format: date-time, nullable: true }
        respondedAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    MarkNotificationsReadRequest:
      type: object
      properties:
        notificationIds:
          type: array
          minItems: 1
          items:
            type: integer
            format: int64
            minimum: 1
      required: [notificationIds]
    NotificationResponse:
      type: object
      properties:
        notificationId: { type: integer, format: int64, nullable: true }
        message: { type: string }
      required: [message]
    NotificationListItemResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        type:
          $ref: "#/components/schemas/NotificationType"
        title: { type: string }
        message: { type: string }
        isRead: { type: boolean }
        createdAt: { type: string, format: date-time }
        readAt: { type: string, format: date-time, nullable: true }
    UnreadNotificationsCountResponse:
      type: object
      properties:
        count: { type: integer, format: int64 }
      required: [count]
...
    PageOfferListItemResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/OfferListItemResponse"
        page: { type: integer }
        size: { type: integer }
        totalElements: { type: integer, format: int64 }
        totalPages: { type: integer }
        last: { type: boolean }
      required: [content, page, size, totalElements, totalPages, last]
    PageNotificationListItemResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/NotificationListItemResponse"
        page: { type: integer }
        size: { type: integer }
        totalElements: { type: integer, format: int64 }
        totalPages: { type: integer }
        last: { type: boolean }
      required: [content, page, size, totalElements, totalPages, last]