# tripmatch_backend/src/main/java/ro/stefanscd/tripmatch/notification/dto/MarkNotificationsReadRequest.java
package ro.stefanscd.tripmatch.notification.dto;

import jakarta.validation.constraints.NotEmpty;
import jakarta.validation.constraints.NotNull;
import jakarta.validation.constraints.Positive;
import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

import java.util.List;

@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
public class MarkNotificationsReadRequest {
    @NotEmpty
    private List<@NotNull @Positive Long> notificationIds;
}