# tripmatch_backend/src/main/java/ro/stefanscd/tripmatch/notification/dto/UnreadNotificationsCountResponse.java
package ro.stefanscd.tripmatch.notification.dto;

import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
public class UnreadNotificationsCountResponse {
    private Long count;
}